requirements.txt
transformers
torch
faiss-cpu
numpy
sentence-transformers
spacy
nltk
pandas
ollama